requirements.txt
asttokens==3.0.0
comm==0.2.2
debugpy==1.8.12
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.2.0
filelock==3.20.0
fsspec==2025.9.0
ipykernel==6.29.5
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.7.2
MarkupSafe==3.0.3
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.6
opencv-contrib-python==4.12.0.88
parso==0.8.4
pillow==12.0.0
prompt_toolkit==3.0.50
psutil==7.0.0
pure_eval==0.2.3
pywin32==308
pyzmq==26.2.1
stack-data==0.6.3
sympy==1.14.0
torch==2.9.0+cu126
torchvision==0.24.0+cu126
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.15.0
wcwidth==0.2.13
